requirements-docker.txt
# Backend runtime dependencies for Docker container

# Core dependencies
langchain>=0.1.0,<1.0.0
langgraph>=0.0.40
openai>=1.3.0
neo4j>=5.0.0
tiktoken>=0.5.0

# FastAPI and API dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Document processing
pypdf>=3.17.0
python-docx>=0.8.11
python-multipart>=0.0.6
python-pptx>=0.6.21
openpyxl>=3.1.0
marker-pdf[full]>=1.10.0

# OCR and image processing
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.17.0
opencv-python>=4.8.0
langdetect>=1.0.9

# Configuration and environment
pydantic>=2.0.0
pydantic-settings>=1.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
httpx>=0.24.0
